satellite/metabase/rangedloop: validate number of segments
The main goal for this change is to compare number of segments processed
by loop and number of segments in DB before and after loop. This
comparison make sense only if all parts are using the same stale read
timestamp. This should give us the answer if the loop is skipping
segments or not.

Implementation is similar to `LiveCountObserver` but for now I wanted
to keep this only for GC bloom filter generation process.

This change has only basic tests, as it would require more work and
I'm not sure if this code will live long.

Change-Id: Ic4d6cfc123adb1dc0141410dd6151ee7690702d8