storagenode/hashstore: raise rewrite probability to a power
this squishes down the probability curve so that
for power values >1, logs are less likely to be
rewritten until they are closer to the alive
fraction. in the limit as the power goes higher,
this approaches the hard cutoff we used to have.
similarly, as the power approaches 0, the
probability flattens out and approaches 1 for all
alive fractions, increasing the eagerness of
rewrites.

Change-Id: I53775b9a6fc4416cab242fcd04fd3211f2c3e23b