cmd/satellite: Add flag for batch size delete objects
Add a flag to the "delete-data delete-objects" satellite command to
specify the batch size of the "delete all the objects of a bucket"
because we observed the Spanner error

```
spanner: code = "InvalidArgument", desc = "The transaction contains too
many mutations. Insert and update operations count with the multiplicity
of the number of columns they affect. For example, inserting values into
one key column and four non-key columns count as five mutations total
for the insert. Delete and delete range operations count as one mutation
regardless of the number of columns affected. The total mutation count
includes any changes to indexes that the transaction generates. Please
reduce the number of writes, or use fewer indexes. (Maximum number:
80000)"
```

when we wanted to delete some bucket's data in the EU1 satellite.

Change-Id: I4b3c8324ab92793a4c4536d975b24f775e85a3e0