satellite/jobq: initial implementation of jobq ("job queue")
This is meant to provide a near drop-in replacement for our
PostgreSQL-based repair queue.

jobq runs as a standalone service (or it can run in-process for
requirements like testplanet). It provides a separate queue for each
placement. It can update jobs already in the queue. All queues are
in-memory only.

Every placement really gets two priority queues: one for ready repair
jobs, ordered by priority; and one for repair jobs that are not yet
ready to retry, ordered by the time when they will be ready. Jobs are
automatically moved from the retry queue to the repair queue as they
become eligible. Jobs inserted will be placed in the appropriate
priority queue based on the LastAttemptedAt field.

Pushes and pops to/from a specific placement are O(log n) amortized.
Updates to an existing job and inspecting queue lengths are O(1). Pops
are also possible from "any of this list of included placements, or not
including this list of excluded placements". Such pops are O(p + log n)
where p is the number of different placements supplied to the command.

After the queue grows to a large size and shrinks again, the memory it
no longer needs can be unmapped on *nix systems, so that memory usage
does not have to go continually up.

A tool called jobqtool is included; this can be used to query the
lengths of job queues for specific placements, or add or remove queues
for placements.

jobq speaks DRPC (either with or without the magic DRPC header) over
TLS. Both sides should have TLS configured so they verify the other side
is as expected.

One extension from the previous repair queue functionality is that it
can keep track of the numbers of missing pieces and pieces out of
placement for each repair job. Those do not solely determine priority,
but they can help give a more human-compatible idea of segment health.

Still planned: /debug endpoints for inspecting queue sizes and segment
health.

Change-Id: I9f1dfcd9484f393e020332dc5198d16e5dc78a44