storagenode/modular: remove support for arm32 docker images temporarily
We're seeing net/http: TLS handshake timeout while
building the arm32 images, and this seems to be a known
issue in the docker community.

We don't need the arm32 images anyway but provided
it for rpi users who need it for testing.
Let's remove it temporarily till we figure out a fix.

Change-Id: I0f9f0324ae99834febddb2e8da863a06c030571b