storagenode/piecestore: migrating backend should use old backend only if piece was missing
MigratingBackend can use both Hashstore and the old Piecestore and the same time.

With `ReadNewFirst=true` settings, it checks the hashstore first, and if piecestore returned with error (!!) it tries to use the old hashstore.

This means that old backend will be used, even if connection is closed or context is cancelled.

Michal/ noticed lot's of failures from old piecestore backend, even on nodes without piecestore data.

This patch improves this algorithm to check the piecestore ONLY if file was missing in hashstore. In case of any other error, we return with the error.

This should be safe, because our testing with hashstore goes well, in case of any error, we probably have a real problems (or just having a long-tail cancelled request).

This is important, as we wouldn't like to touch blobstore if it's not necessary, because old piecestore may have huge directories, which can be read only very slowly.

The current implementation may generate cascading failures (slow node --> cancelled connections --> more blobstore read --> even slower nodes).

It's not very easy to test this change, but I created a custom DRPC client and cancelled the client after 1 second.

When I used debug (to make sure I cancel the connection during the initialization of the code), I was able to reproduce the problem.

Change-Id: Ib2a38e8df3dc3dcf1f7aa369cb8a478e3d6aef67