storagenode/hashstore: initial memtbl implementation
there's still lots of work to do, but this
implements the basics of the memtbl.

the memtbl is a replacement for the hashtbl that
uses memory to make the disk writes linear instead
of random. they both provide an interface that is
effectively a map[Key]Record for the Store to use
and the hashtbl does this by backing all the reads
and writes with a file on disk.

instead, the memtbl just directly keeps the
map[Key]Record in memory but with an append only
log of updates to that map that we read in during
opening for persistence. unfortunately, this takes
too much memory at an expected 96 bytes per entry.

the first insight is that the Record includes the
Key, so we can change it to a map[Key]Value where
Value is the fields of Record without the Key.
This goes from 96 bytes to 64 bytes.

the next insight is that we don't need to keep the
values in memory, we can look them up from the log
file on reads and just store the offset of the
value in memory. this would be a map[Key]uint64
going from 64 bytes to 40 bytes.

the next insight is that we don't need the exact
offset of the values since they are all fixed
width, so if we keep the writes aligned we can
store the index instead. this goes to a
map[Key]uint32 going from 40 to 36 bytes.

the next insight is that we don't need to store
the full Key which is 32 bytes. the Keys are
cryptographically random and so the probability of
a collision with 12 bytes would be neglibile and
we can store the first 12 bytes of the Key in the
map and ignore collisions. this goes from 36 to 16
bytes.

the next insight is that we can actually handle
collisions nicely if we keep two maps and reserve
an offset to signal that there was a collision.
this lets us drastically reduce the amount of Key
we need to store because we can allow a low but
not non-neglibile probability of a collision, and
reduces the map to a map[[4]byte]uint32 going from
16 to 8 bytes.

the final insight is that [4]byte is too small and
for large numbers of entries that we expect, and
the number of collisions actually pushes us to use
up to 10 bytes per entry on average, but going to
a [5]byte is a proble due to the way entries are
laid out internally in go maps. a uint32 value
requires an alignment of 4 bytes, so using [5]byte
means that interally each entry requires 12 bytes.
if instead we store the value as a [4]byte, the
alignment becomes 1 byte so they are densely laid
out leaving us with an avergae of 9 bytes per
entry at all expected lengths of the map. see the
comment in map_test.go for a table showing these
sizes.

thus, we arrive at the final design of the memory
table: a map[[5]byte][4]byte for the common case
of no collisions with a map[Key][4]byte to resolve
the collisions, and a file handle that we append
the inserts into to reconstruct these maps.

early benchmarks against an SSD show sigificant
wins in inserts and compaction:

           │   hashtbl   │               memtbl |
           │  keys/sec   │  keys/sec    vs base |
Lookup/14    769.6k ± 0%   793.6k ± 1%   +3.12%
Lookup/15    767.1k ± 0%   785.2k ± 1%   +2.36%
Lookup/16    762.9k ± 0%   775.3k ± 1%   +1.62%
Lookup/17    745.7k ± 2%   741.4k ± 1%        ~
Lookup/18    711.4k ± 1%   713.8k ± 1%        ~
Lookup/19    692.3k ± 1%   685.5k ± 1%   -0.97%
Insert/14    329.6k ± 0%   484.1k ± 0%  +46.87%
Insert/15    340.4k ± 1%   487.7k ± 0%  +43.26%
Insert/16    345.1k ± 1%   487.1k ± 0%  +41.12%
Insert/17    345.7k ± 1%   483.2k ± 1%  +39.79%
Insert/18    343.9k ± 0%   483.3k ± 1%  +40.53%
Insert/19    340.3k ± 0%   479.4k ± 1%  +40.89%
Compact/14   1.176M ± 5%   1.967M ± 2%  +67.20%
Compact/15   1.366M ± 6%   2.020M ± 1%  +47.81%
Compact/16   1.479M ± 3%   1.966M ± 2%  +32.91%
Compact/17   1.529M ± 2%   1.923M ± 2%  +25.75%
Compact/18   1.565M ± 3%   1.872M ± 2%  +19.60%
Compact/19   1.595M ± 3%   1.829M ± 1%  +14.68%
geomean          714.4k        887.2k   +24.18%

Change-Id: I3d7683ba81313d89e24e3cc295bca9d39432ed6b