storagenode/piecestore: fix error labeling in case of cancellation
The existing code tries to replace any kind of error with context-canceled
(in case of context cancellation) in defer group.

But it doesn't work, as rpcstatus.Code(err) != rpcstatus.Canceled is always false.

It's better to check if either the name or the rpc status is wrong.

Tested with 10-12 hours of production traffic and classification was fixed.

Change-Id: Ide991159b31a2c1032b3c517cd13c30885c201a1