satellite/nodeselection: support dedicated upload filters
There are cases when we need additional filters for uploads.

It means that we are fine with the already uploaded data, but newly updated data lands only to selected nodes.

Until now, we did this with selectors. We implemented same filtering capabilities to each selector.

Example:

```
filter: tag("...","best-of-my-nodes","true")
selector: weighted("tag:$SIGNER_ZERO/weight", 100.0, successfulAtLeastPercent(uploadFailureTracker, 0.33) && exclude(tag("...","maintenance","true")))
```

This is good, but it makes very hard to create analyitics.

Which nodes are included in this placement? How many nodes are in the this placement? And how many free space?

`best-of-my-nodes` are included, but the rest of the filter is included in selector, which makes it impossible to use.

This patch is a small modification to make it possible to define filters (1) outside of selectors, (2) only for uploads.

```
filter: tag("...","best-of-my-nodes","true")
upload-filter: exclude(tag("...","maintenance","true"))
selector: weighted("tag:$SIGNER_ZERO/weight", 100.0, successfulAtLeastPercent(uploadFailureTracker, 0.33))
```

With this approach, we can easily do additional metrics, like available free_space per placement.

And -- I believe -- it's also easier to read.

Change-Id: I62027828524ad83e613f7fd5979b9a3f9919da38