satellite/satellitedb/projectaccounting: count the tally data from across-month boundaries
consider two months, for example say February and March, that
we are generating invoices for using this function.

the code, prior to this change, will query the database for
all tallies that have interval_start in february, and none in
january or march, which seems reasonable.

the problem with this strategy is that each tally is actually
a fencepost around a range of time, and we need to figure out
how many hours elapsed within that range. by only considering
the fenceposts in february, we never look at the span that
crosses from january into february, or the span that crosses
from february into march.

we also don't want to count these spans twice, so we don't
want to count both inclusive in february.

so we're going to pick one - we picked the span from january
into february, and include that span in the february range.

some considerations:
 * we currently process tallies in decreasing time order
   so newest to oldest. this means that for a growing
   bucket, we will undercount the amount of byte hours
   stored in the bucket. if we processed instead in
   increasing order, we would overcount the amount of
   byte hours actually stored in the bucket.
   undercounting seems more fair, and is likely why we
   did it this way. but it's worth additional
   consideration
 * we could choose the other span, the february into
   march one. this would have some nice benefits, like
   making it so all the tallies we consider were actually
   calculated in february, but it has the downside of
   delaying a small amount of revenue another month.

This commit include the fix for not missing any period on our
monthly invoicing process and add a test that verifies the
storage calculations across several months.

Change-Id: Ifbdee6d52f81788ef39f88cb243729a29e7a7d9c