storagenode/hashstore: revive at the db layer
this change makes it so that we check both stores
for a non-trash piece before we try to revive so
that we don't end up with spurious restore from
trash events. this can happen because we have two
stores and sometimes pieces get uploaded to the
same node twice. thus it can be that it gets
uploaded once, flagged as trash (maybe long tail
cancel, maybe repaired off the node), and then
uploaded again. then reads can end up reading from
the store where the piece is trash when it's not.

Change-Id: I060914002ae8b66fd1ff1cf6883bb62c0740fbd2