satellite/nodeselection: split-based topology selector parameters
This is a variation of JT's invention from https://review.dev.storj.io/c/storj/storj/+/16360

Current topology selector can be configured as [3,2,1], [tag:datacenter, tag:server].

Which means that it will:
  * select 3 datacenters
  * from each datacenter, 2 servers
  * from each server 1 instance

3 * 2 * 1 = 6 selected nodes all together.

This is very clear, but it uses the hard-coded number of selected nodes (3 * 2 * 1), instead of the requested one. In case of a user requests 20 nodes, it will fail, as only 6 is selected based on topology rules.

JT's suggestion was to use split based approach, where `[3,2,*]` means:

 * select 3 datacenters, and from each datacenter try to get `requested` / 3 nodes
 * inside datacenters, pick max 2 servers
 * from each server, pick `requested` / 3 / 2 instances

An interesting behavior of this approach is the possibility of using very high numers.

For example with `[3,1024,*]`, we will try to select 1024 servers per datacenters. Obviously, we can select only as many as available, so we try to select all.

But because we calculate the required number of selected nodes, based on the real selected ones, we will always select minimal number of instances from the maximum number of groups.

There is one difference between this patch and 16360: this calculates the required amount in a different way: it's more dynamic.

If the first few groups doesn't have enough servers (or enough instances), the next groups will be called with more required nodes.

This may break some invariants (which are softer than the original, hardcoded numbers, anyway), but probably more resilient to the situation, when high number of servers (but not all) are stopped in a datacenter.

Change-Id: If42c839c08780388189438641d8806a81b1e3037