satellite/repair: define queue.Release method
At the meeting point of two different paradigms (sql-based priority
queue and heap-based priority queue) lies this conundrum: after you
select a job, should you have to call something if the repair _fails_ or
if the repair _succeeds_?

With the sql-based queue, due to the nature of the implementation, we
had to update LastAttemptedAt when a job was selected, and require
another call (Delete) if that repair succeeded, to remove the job from
the queue.

With the heap-based queue, however, when we select the highest-priority
(lowest-health) job, we have to remove that job from the queue. So in
this case, we require another call (Insert) to put the job back in the
queue in the event the repair _failed_.

To facilitate both of these models coexisting, we add here a method
called Release, which should be called after every repair finishes,
whether the repair succeeded or failed. Success or failure is passed in
to the Release call, and the underlying implementation decides what to
do about that result.

Change-Id: Id667d124e2c7c547df50260380eaca8cce13e5ef