satellite/jobq: make Inspect() indicate whether segment found
Previously, Inspect would return a job object with a zero stream ID if
the indicated job was not found. Rather than relying on that clunky
approach, it will now return a boolean indicating whether the segment
was found. We can use that cleanly in the jobq client library to return
ErrJobNotFound in that case.

Change-Id: I3566237ecd39d44c04c04d20d06fa7daafa5f078