satellite/jobq: ServerTimeNow sentinel value
It will be helpful in some tests (and, generally, in real usage) to say
"fill in LastAttemptedAt or InsertedAt with the server's current time".
This magic sentinel value (a valid Unix timestamp and a valid time.Time,
but not one that we expect to see ever used for those fields) instructs
the server to replace it with the current server time.

is mainly important for cases of clock mismatch. In cases where the
server clock differs from the client clock, it is important that newly
set LastAttemptedAt values use server time and not client time. We don't
want to use a zero/unset time to indicate "fill in server time" because
a zero/unset time already has a "null" meaning; i.e., a job that has
never had a repair attempted yet.

InsertedAt is less critical to proper functioning, but for the sake of
consistency we should use server time there too.

UpdatedAt is always set by the server on insert or update regardless,
so ServerTimeNow is not relevant there.

Change-Id: I4ed823980acdc11a2d12b119ad84cf998e2e186d