satellite/metabase: fix requerying with delete markers
This fixes requery in the presence of many delete markers.

Previously the querying assumed that there never is a completely
empty batch returned from the spanner. This is however not true
when there's a delete marker and many "deleted objects".
So, when we detect a delete marker in a batch we'll just
bump the limit, because we don't know that in advance.

Change-Id: Ia8fb159a1d03821f6ee69d93ed0435ffe2f042d4