cmd/tools/metabase-minimize-listing-csv: tool for shrinking data
For metabase listing bugs it's happened several times where we get a
large set of encrypted object keys as the faulty input. This tool
minimizes this input and replaces the paths with incrementing numbers.
This makes the input smaller and also easier to spot what's going on
with the input.

Change-Id: I3f7c6b87fa0fcd9660c44214c105fb2807965ea1