cmd/satellite: Accept an email to delete data of an account
Modify the satellite delete-data sub-commands to accept, apart of path
to a CSV file, a single email address to ease the process to delete the
data and account of a single user.

Deleting the data and the account of a user is a usual use case of the
production owner. Without this, the production owner has to create a CSV
file only for that account. With this changes, the production owner only
has to pass the email address a positional parameter.

Change-Id: Ide7686c81d40faa05013a8092e0ea435b1450438