web/satellite: improve the share dialogs
Improve the explanation of the shared links, including additional information about each link. Add icons. Organize the dialogs with tabs, links, social, and embed. Use one line input fields for the links.

Update copy link button to be based on the active tab, and show a chip of the default link which is copied if there is also a raw direct link available.