storagenode/hashstore: lazy+sem acquire log files
this changes the way we write data into the store
to buffer the data in memory and then acquire a
log file at time of commit instead of acquiring a
log file right away and writing directly to it.
additionally it adds a semaphore to the commit
phase to limit the number of concurrent writes we
do. this should trade off some memory usage for
less fragmented writes to the log files which may
help with disk throughput. additionally, it makes
it so that canceled pieces don't have to do any
writes to disk, reducing overall iops.

Change-Id: If0c00a34f6a009fbe469db85608f3496762a4ce7