storagenode/hashstore: mmap support
by setting the STORJ_HASHSTORE_HASHTBL_MMAP env
var to true, the hashtbl will use mmap for reads
and writes which avoids copying the data into
memory before flushing and may perform better with
the kernel's page cache.

it remains to be seen if this will interact poorly
with the scheduler for pause times but recent i
saw https://www.youtube.com/watch?v=1I1WmeSjRSw
and it convinced me it should be fine.

most of the code motion is refactoring the
platform specific code into a platform package
because it grew yet another thing, so it seemed
like it was time. because the hashtbl already did
all of its record reads/writes through some cache
interface, adding an mmap based one was natural
and easy.

unfortunately the tests don't run with mmap mode
by default. i think some future work is required
to figure out a good way to run the tests with all
the different configuration options that are being
exposed. for now, it sufficies for me that it does
pass locally and on wine.

Change-Id: I1c79d1085ac6f252227b9bfeea4a30a045ef5c04