satellite,private/testplanet: exercise jobq in tests
We don't want to add an extra test for jobq where it's not necessary. We
also don't want to stop testing the sql-based repair queue (for now). To
meet these needs we introduce here a testplanet flag 'ExerciseJobq'. If
set, a testplanet test will be run an extra time using jobq as the
repair queue.

ExerciseJobq has been added to all the tests that materially use the
repair queue.

A few tests needed tweaking to be able to work with the lower-precision
timestamps used in jobq.

Change-Id: I71c80234a64cf27c5528e8ddebdb550b8b43eab2