satellite/metainfo: help de-flake TestEndpoint_checkRate
One cause of problems was that the RateLimiter was managed by an LRU
cache, which would expire after 10s and provide a whole new RateLimiter.
So anytime the test took more than 10s it would break.

Another problem that revealed itself later was that requests are also
subject to the DownloadLimiter (which isn't being tested here).

Hopefully with these out of the way the test will succeed more reliably.

Change-Id: I2ab5bf129a2102542aeae724d0f5a646ceccf174