satellite/jobq: switch to a minmax heap
If we want to drop the lowest-priority repair jobs when a maximum
capacity is reached, we will need a double-ended priority queue.

This will allow O(1) identification of the lowest priority item in much
the same way we can currently identify the highest priority in O(1).
Removal of either one is now O(log n).

Change-Id: Icee726af8b616241e0739a4bfe0c091f4055d48d