storagenode/hashstore: use constructors for new tables
instead of the ExpectOrdered api, we introduce the
concept of a constructor where you append entries
to a table, hopefully in an efficient order. the
constructor can do appropriate buffering. this
lets us have "all or nothing" semantics which is
really what we want for the memtbl and lets us
avoid worring about tricky cases where we have
some partial writes that we then have to undo
which requires deletes from the in memory map,
which complicates a future mmap based
implementation.

Change-Id: I5155fa8d8ceac2398db94b457dde29ffe2b4ee4a