shared/dbutil/spannerutil: avoid leaking admin clients by reusing existing
This change prevents the creation of database and instance admin
clients for the Spanner emulator when ones are already present.
Previously, the methods responsible for ensuring these clients'
existence would create new ones instead of preserving those that were
constructed before. This caused a leak in the clients' underlying
resources, particularly gRPC connections.

Change-Id: I76ef67de01f3b7a4e0a99f32497ab0f1c298d1d3