cmd/satellite: Add new sub-command set status user accounts
Add a new sub-command to set a specific status to a single or a list or
user accounts.

I created a new command and moved 2 other sub-commands under the
`delete-data`  to this one, resulting in dropping the `delete-data`
command. The new sub-command is under this new command. This new command
(`users`) is more general than the previous and scope well the 3
sub-commands.

The commit changed the method of the private type that reads CSV to not
ignore accounts with a specific status because the new sub-command
doesn't do any check of the current status. This provoked to add the
status verification into the two existing functions that contains the
majority of the logic of the existing sub-commands. That method is
renamed to create a new on with the name of the existing one to only
return the users without projects because the new sub-command doesn't
need the projects.

There are minimal updates in the documentation of the methods to delete
objects and accounts to reflect that the "inactive" accounts are skipped
and logged.

Implemented the methods required by the pflag.Value interface to the
UserStatus type because the new command requires to receive it as a flag
parameter.

The commit implements the new method to set the accounts status leaning
on the type to read the CSV data and following the same structure used
by the existing delete-data functions.

Change-Id: I150903ed1b9d77707520e128da407e41d5b3f449