storagenode/hashstore: use custom map type
implements a map type based on swissmaps that only
has a single buffer (gc friendly, mmap friendly)
and optimized for our use case where we only have
to do a single lookup to do an insert or update
because it returns a pointer to the slot that
either contains the value or would be written to.

a table is a collection of groups, where each
group contains 8 entries, has an 8 byte header
with 1 byte per entry followed by 72 bytes of 9
byte entries. each header byte is composed of a
single bit flag indicating if the corresponding
entry is set followed by a 7 bits fingerprint of
the hash of the entry for quick comparisons.

the idea is that you can find a group and quickly
check all 8 entries by doing some 64-bit bit-wise
math to generate a bitmask of the entries that are
set and have matching fingerpints, giving a false
positive of about ~1/127. similarly, it's easy to
find the slot in a group you should insert into by
doing a popcount on the header.

Change-Id: Ibea3a82d6fe946cfcf5acde20e15f9bd44f248d3