storagenode/hashstore: make db.Compact only compact passive
compacting the active store is dangerous because
it can race with an incoming Create and that can
lead to it blocking until the compaction
completes. worse, the same lock is held for monkit
stats and so oopsie that can be pretty bad.
Compact is only ever called in tests (or from the
wacky crawlspace server) so this doesn't really
happen but this moves it towards being something
we can call as an endpoint on nodes.

Change-Id: I1fd5222ae3e6080ea099b4a90f0531cc4f0fa2a2