storagenode/hashstore: mmap for memtbl
this commit adds an mmap mode for the memtbl by
using mremap to extend the size of the mapping
whenever there's an extra page of records. this
may seem too frequent but the benchmarks in a
fully cpu bound workload demonstrate that insert
performance is barely affected (~5% ish, or ~100ns
per key) because we only remap every 64 inserts.
in a real workload, it will be totally dominated
by a single disk seek. compaction is unaffected by
it because during the constructor we don't bother
with mremap because we expect to not have to
lookup, so we can just remap once at the end.

because of the advanced mmap usage, we drop
support for mmap on any platform besides linux. we
can consider adding it back in later when we have
a better idea of all the features we'll require.

because we support only linux, we add code to
madvise the mmap to sequential when doing range
for both the hashtbl and memtbl and convert it
back to random when we're done. hopefully this
will help the kernel prefetch pages better.

Change-Id: I3cd5c3fb2be900fd25d70909b2f5a2cf978f9ae5