storagenode/hashstore: mlock mmap by default
if we're mmaping, mlocking the pages causes them
to be kept resident. this can be useful if you
are sure you have enough memory for the index
structure because the kernel is more than happy
to evict them which is usually silly.

also, the api of returning a close function was
useful for the windows mmap call because it needs
more than just the byte slice to do a close but
it's a problem for remapping on linux because it
would try to unmap closed over old pages. now that
we don't have windows support we can do the easy
thing and circle back when we need to.

Change-Id: I44e8d6ae2feb07036f6a5c1b6c59e31268e40e14