satellite/metabase: fix deletion of previous object version
Some time ago, we changed this method to not try to delete previous
version if no object under this location was found. The issue with
this change was that we missed that at that point in time we always
have pending object there as we just started upload.

This change add additional condition that checks if object we found
have version 1 (first possible) and is a pending object. That gives
us information that we don't have to try to delete previous version of
object under this location.

Change-Id: I2af70f55fb871cd7822803298bfe4a776395042d