storagenode/hashstore: rewrite records ordered by log/offset
this change first collects an approximation of the
records we want to rewrite (it can rewrite more if
some complicated computations cause the record to
become expired during the real hashtbl rewrite,
and i don't think it can be under but it would
still be correct if it was), maintaining an index
that first sorts by the log/offset and rewrites
them all to minimize seeks, and then sorts by key
for O(log n) lookups when rewriting the hashtbl to
see if it was already done.

this structure is just a flat list of records that
we sort multiple times to keep memory usage low as
it requires only storing the record itself. we
usually only rewrite a couple hundred thousand
records per compaction but even at a million that
would be 64MB of memory only during compaction
which seems acceptable.

there's an environment variable to disable this
called STORJ_HASHSTORE_COMPACTION_ORDERED_REWRITE
which is default off, but on for tests.

Change-Id: I49f34aae1144bee2d0ff6f6027b9bcda9203a175