storagenode/hashstore: make rewriteRecord copy_file_range again
i dunno when this optimization was lost, but the
writer having manual and automatic modes was
complicated and causing problems. adding a
ReadFrom method to it would allow copy_file_range
to kick in but that was super complicated with the
potential for having a file handle or a memory
buffer.

anyway, turns out that the only user of a manual
writer was rewriteRecord, and so we can instead
avoid the writer abstraction entirely and directly
do an io.Copy from the two file handles which
means we don't have to worry about the stdlib
missing the optimization. there's a little bit of
duplication of code in book keeping the file
handle size estimates and appending the record for
log recovery, but that seems worth it.

we also are now sure that the property that
rewriting an entry only updates the log and offset
fields of the record holds because the code
doesn't pull a rec from the writer and hope.

Change-Id: Ie088cbe4047e02b1a95b7158c7eb610c36eb4846