storagenode/hashstore: make stats a little better
we can load the data/logs rewritten live instead
of waiting for compaction to finish to give
more timely stats.

also most of the time compacting is during
rewriting so we can reuse the total/processed
records values to show the two phases.

adds a data reclaimed metric that can be compared
to data rewritten to get an idea of the cost of
compactions (how many bytes written to reclaim).

Change-Id: I2dc2f3c333cd49b37f60293a6a22f325fd42ed47