satellite/jobq: fix potential crash
When the repair queue is entirely full and the retry queue is entirely
empty and we try to insert into the retry queue (or the other way
around), a crash can result as we try to remove items from the empty
retry queue to make "room" for the new one (since the max size applies
to lenRepair + lenRetry).

With this change, when the maximum number of items is reached and we
need to insert more, we pop from either the repair queue or the retry
queue; whichever has more items at the time.

Change-Id: Ib11f6f79e338fd75f53f3be5e9e7b17d727852db