storagenode/hashstore: better defaults
many nodes on the public network are spending too
long compacting because once they have a large
size of data, some log is always a candidate for
rewriting and so it spins rewriting a log over and
over.

blunt this by making the multiple higher (so that
you need 10x as much data before you start to get
to the point where you're always rewriting) and
increase the probability power to 2 (so that you
need way more data before you hit the point where
every compaction observes a log to rewrite).

Change-Id: Iceae2bbf51702a6e8c5eb5956f3f634e63c6ebc7