satellite/accountfreeze: fix early final warning email
This change fixes an issue where a user was sent their final freeze
warning email earlier than expected. The issue happened because the
email intervals were being compared to the date the freeze event was
created instead of the date the previous email was sent. The configured
intervals are now expected to be cumulative to resolve this.

Issue: https://github.com/storj/storj-private/issues/1240

Change-Id: I8a965def9d78a880bd03d0d20038596cdfc0cdbf