satellite/payments/stripe: Log error before retry
Log the error reported by Stripe as warning before retry the request.

Without this log, we may miss what was the initial error because
subsequent requests executed by the following retires may return a
different error.

I think that this is a problem that we got and we can only see the last
error that was related to duplicated idempotency key.

Change-Id: If6d6451f0b0983661d1db64b1e9eddaa69989d14