cmd/satellite: Replace user-accounts set-status cmd
Replace the `satellite user-accounts set-status` command which was
mostly used to set accounts to `pending deletion` status by a safer
implementation: `set-status-pending-deletion`.

The commit implements this new command and its test and deletes the
previous `set-status` command and everything related to it.

The new command only set accounts to `pending deletion` status and only
when the accounts:
- Are in `active` status
- Are NOT in the `paid tier
- Have a `free trial expiration` freeze
- Is not a member of a third party project

The new command check before the update to have a more comprehensive log
message, but it's safe to proceed with the update because a new method
implemented in the consoledb users repository only sets the user status to
"pending deletion" atomically making the verifications.

Related issue: https://github.com/storj/storj/issues/7446

Change-Id: I8d1ef1e830096f87080649bc86827f9506eb15c7