satellite/metabase: try using mutations while committing segment
To use mutation to insert or update segment into DB we need to split
operation into read and write parts. Read is checking if parent object
exists, write is inserting segment into DB.  It's not certain that it will give use better throughput so this feature is behind feature flag
for easy testing.

Change-Id: I47c01c5132e5e92ebcda370b84ae321270ad7434