satellite/metainfo: update attribution logic to store placement
When bucket is created, insert bucket placement into attribution DB. If
attribution already exists for bucket, if placement in attribution is
different than that of the bucket being created, return error.

Attribution is also checked/set during uploads, and the logic there has
been changed a bit. It used to be that attribution was only inserted if
user agent was not nil, since recording the user agent was the purpose
of the table. There is a soft check on attribution during an object
upload, which just checks if a row already exists. Now that we insert
attribution during bucket creation, even if no user agent is set, the
soft check instead checks if the user agent is null. If so, the
attribution is updated with the user agent.

issue: https://github.com/storj/storj-private/issues/1235

Change-Id: I7eb4894605619bd4add5a4c2b8a99ab98a34bf69