web/satellite: show link expiration in share dialog
This change displays the expiration date of share links in the Share
dialog.

The authservice may impose limits on the expiration date of access
grants registered by users in a free trial. When the access grant
associated with a share link expires, the link will become unusable.

References storj/storj-private#1139

Change-Id: I5e3b9b485984a7e21bffd37390466195f4f6fea2