satellite/satellitedb: account for completely deleted buckets
this code tries to make sure we make a final zero in the bucket
storage tallies table when a bucket is emptied. unfortunately,
the old query relied on bucket_metainfos, where the row there
is also deleted if the bucket is deleted!

this changes the query to be able to zero out deleted buckets
too.

Change-Id: Idccf989711e29d45d8faaeb089cf9e01d0dd1101