storagenode/hashstore: compaction only waits for flush
writers started buffering in memory and only
getting a log file and flushing to disk on close
so there's no reason for an open writer to block
a compaction. instead, we can have compaction
only wait for no flushes to be in progress. this
should be a much smaller amount of time and so
we won't have compactions waiting up to minutes
to do a couple ms of work.

Change-Id: I6f2dd7c663b577aa4b55145928cddaeef3c1345d